#15 Initialize basic data for addresses, accountHolders, admins and thirsParties

+++ src/main/java/com/proyecto/abanca/dto/CheckingDto.java
@@ -12,7 +12,8 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CheckingDto {
-    private BigDecimal balance;
+    private Long balance;
     private String primaryOwnerId;
     private String secretKey;
+
 }
